[WIP] Adding support for syncthing local discovery
This code should be almost ready. A few questions remain in the code,
at least two are of importance:

1. I modified the wsutil/base32.* files, but a function I moved wis
referenced in libwsutil0.symbols. I guess I need to do something about
that, but what?

2. I have a broken filter, when you select and apply one of the
top-level trees (see line 387 in packet-syncthing.c), and I don't know
how to solve the issue. Any hint is welcome.

Concerning the modifications in the base32.c and .h files, these are
related to a mail I sent to the dev mailing list on Nov 29, 2018.
I won't repeat the details here.

In a few words, it turned out that the function that was called ws_base32_decode was in fact an encode, and was also not the standard base32. Since I had no negative reply from the mailing list, I assumed my interpretation was correct and went ahead with the changes.

In base32.c, we now have a ws_base32_encode function.

I'll be running fuzz tests tonight and remove the todo if nothing
crashes.

Bug: 15283
Change-Id: I5838cf78c28fc57f305562debed91841040ec193